Minor bugs fixes and docker support.

+++ auth/src/main/java/lk/chethana/security/repository/ApplicationUserRepository.java
@@ -3,7 +3,7 @@ package lk.chethana.security.repository;
 import lk.chethana.security.entity.ApplicationUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplicationUserRepository extends JpaRepository<ApplicationUser, Long> {
+public interface ApplicationUserRepository extends JpaRepository<ApplicationUser, Integer> {
     ApplicationUser findByUsername(String username);
     ApplicationUser findAllByUsernameAndPassword(String username, String password);
 }
