+++ auth/src/main/java/lk/chethana/security/repository/ApplicationUserRoleRepository.java
@@ -3,6 +3,6 @@ package lk.chethana.security.repository;
 import lk.chethana.security.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplicationUserRoleRepository extends JpaRepository<Role, Long> {
+public interface ApplicationUserRoleRepository extends JpaRepository<Role, Integer> {
 
 }
